refactor mobile import ServerConfig
